Auth implemented

+++ src/services/cat.service.ts
@@ -1,36 +1,47 @@
 import { Cat, CatDocument, ICat } from "../models/cat.model";
 
-const saveCat = async ({ name, breed, image }: ICat): Promise<CatDocument> => {
+const saveCat = async (
+  { name, breed, image }: Omit<ICat, "user">,
+  userId: string
+): Promise<CatDocument> => {
   const cat = new Cat({
     name,
     breed,
     image,
+    user: userId,
   });
   await cat.save();
   return cat;
 };
 
-const findCats = async (): Promise<CatDocument[]> => {
-  const cats = await Cat.find();
+const findCats = async (userId: string): Promise<CatDocument[]> => {
+  const cats = await Cat.find({ user: userId });
   return cats;
 };
 
-const findCatById = async (id: string): Promise<CatDocument | null> => {
-  const cat = await Cat.findById(id);
+const findCatById = async (
+  id: string,
+  userId: string
+): Promise<CatDocument | null> => {
+  const cat = await Cat.findOne({ _id: id, user: userId });
   return cat;
 };
 
-const findByIdAndDelete = async (id: string): Promise<CatDocument | null> => {
-  const cat = await Cat.findByIdAndDelete(id);
+const findByIdAndDelete = async (
+  id: string,
+  userId: string
+): Promise<CatDocument | null> => {
+  const cat = await Cat.findByIdAndDelete(id, { user: userId });
   return cat;
 };
 
 const findByIdAndUpdate = async (
   id: string,
-  { name, breed, image }: Partial<ICat>
+  userId: string,
+  { name, breed, image }: Partial<Omit<ICat, "user">>
 ): Promise<CatDocument | null> => {
-  const updatedCat = await Cat.findByIdAndUpdate(
-    id,
+  const updatedCat = await Cat.findOneAndUpdate(
+    { _id: id, user: userId },
     { $set: { name, breed, image } },
     {
       new: true,
